+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+### Deprecated
+- Deprecated all color constants
+
 ### Fixed
 - DLL loader could fail to load `SDL2.dll` when other tcod namespace packages were installed.